# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Dondok API
  version: 0.10.0
  description: 돈독 프론트엔드와 백엔드 사이의 HTTP 계약
  license:
    name: All rights reserved
    identifier: LicenseRef-Proprietary
servers:
  - url: /
security: []
paths:
...
  /api/asset-types:
    get:
      summary: 현재 가계부의 활성 자산 유형 목록
      operationId: getAssetTypes
      tags: [Asset]
      security: [{ cookieAuth: [] }]
      responses:
        "200":
          description: 가계부마다 제공되는 고정 시스템 자산 종류
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/AssetType" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/assets:
    get:
      summary: 현재 가계부의 상태별 자산 목록
      operationId: getAssets
      tags: [Asset]
      security: [{ cookieAuth: [] }]
      parameters:
        - name: status
          in: query
          required: false
          description: 기본은 신규 거래에 사용할 수 있는 활성 자산만 반환한다.
          schema:
            type: string
            enum: [ACTIVE, ARCHIVED, ALL]
            default: ACTIVE
      responses:
        "200":
          description: 상태 조건에 맞는 자산, 현재 잔액, 종류 코드와 카드 결제 예정액
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Asset" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    post:
      summary: 자산과 최초 잔액 원장 등록
      operationId: createAsset
      tags: [Asset]
      security: [{ cookieAuth: [] }]
      parameters:
        - { $ref: "#/components/parameters/CsrfHeader" }
        - { $ref: "#/components/parameters/IdempotencyKey" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateAssetRequest" }
      responses:
        "201":
          description: 생성된 자산의 authoritative 상태
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Asset" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/assets/{assetId}:
    parameters:
      - name: assetId
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      summary: 현재 가계부의 활성 또는 보관 자산 상세
      operationId: getAsset
      tags: [Asset]
      security: [{ cookieAuth: [] }]
      responses:
        "200":
          description: 자산 상세와 현재 잔액
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Asset" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      summary: expected version 기준 자산·최초 잔액·카드 설정 전체 수정
      operationId: updateAsset
      tags: [Asset]
      security: [{ cookieAuth: [] }]
      parameters: [{ $ref: "#/components/parameters/CsrfHeader" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateAssetRequest" }
      responses:
        "200":
          description: 수정된 자산의 authoritative 상태
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Asset" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "412": { $ref: "#/components/responses/PreconditionFailed" }
    delete:
      summary: 최신 정리 미리보기 기준 자산 완전 삭제 또는 보관
      description: 거래 이력과 보존해야 할 비활성 설정 참조가 모두 없으면 완전 삭제하고, 하나라도 있으면 보관한다. 활성 연결 계좌로 사용 중이면 409로 거부한다.
      operationId: removeAsset
      tags: [Asset]
      security: [{ cookieAuth: [] }]
      parameters:
        - { $ref: "#/components/parameters/CsrfHeader" }
        - name: expectedVersion
          in: query
          required: true
          schema: { type: integer, format: int64, minimum: 0 }
        - name: previewToken
          in: query
          required: true
          description: 직전 정리 미리보기의 이력·잔액·연결 상태를 확인하는 token
          schema: { type: string, minLength: 64, maxLength: 64 }
      responses:
        "200":
          description: 서버가 실제 적용한 완전 삭제 또는 보관 결과
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AssetRemovalResult" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "412": { $ref: "#/components/responses/PreconditionFailed" }
  /api/assets/{assetId}/removal-preview:
    parameters:
      - name: assetId
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      summary: 자산 정리 결과·경고·차단 연결 미리보기
      operationId: previewAssetRemoval
      tags: [Asset]
      security: [{ cookieAuth: [] }]
      responses:
        "200":
          description: 현재 authoritative 상태로 계산한 정리 미리보기
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AssetRemovalPreview" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: DONDOK_SESSION
  parameters:
    CsrfHeader:
      name: X-CSRF-TOKEN
      in: header
      required: true
      schema: { type: string }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: 같은 사용자의 동일 키·동일 payload 재시도는 기존 리소스의 동일 명령 결과를 반환한다.
      schema: { type: string, minLength: 1, maxLength: 100 }
  schemas:
...
    AssetBehavior:
      type: string
      enum: [STANDARD, CREDIT_CARD, DEBIT_CARD, SAVINGS]
    AssetOwnershipScope:
      type: string
      enum: [PERSONAL, JOINT]
    AssetType:
      type: object
      description: 사용자가 추가하거나 이름을 바꿀 수 없는 고정 시스템 자산 종류
      required: [assetTypeId, systemCode, name, behavior, paymentSourceCapable]
      properties:
        assetTypeId: { type: string, format: uuid }
        systemCode:
          type: string
          enum: [CASH, BANK, CREDIT_CARD, DEBIT_CARD, SAVINGS, INVESTMENT, OVERDRAFT, LOAN, INSURANCE, OTHER]
        name: { type: string }
        behavior: { $ref: "#/components/schemas/AssetBehavior" }
        paymentSourceCapable: { type: boolean }
    CardSettingsInput:
      type: object
      description: 신용카드 생성·수정에서 필수이며 settlementAssetId는 자동 정산 사용 여부와 무관하게 필요하다.
      required: [statementClosingDay, paymentDay, paymentMonthOffset, settlementAssetId, autoSettlementEnabled]
      properties:
        statementClosingDay: { type: integer, minimum: 1, maximum: 31 }
        paymentDay: { type: integer, minimum: 1, maximum: 31 }
        paymentMonthOffset: { type: integer, minimum: 0, maximum: 2 }
        settlementAssetId: { type: string, format: uuid }
        autoSettlementEnabled: { type: boolean }
    DebitCardSettingsInput:
      type: object
      description: 체크카드 생성·수정에서 필수이며 지출 posting이 차감될 결제 계좌를 지정한다.
      required: [paymentAssetId]
      properties:
        paymentAssetId: { type: string, format: uuid }
    SavingsSettingsInput:
      type: object
      description: 적금 생성·수정에서 필수인 자동이체 출금 계좌와 달력일 설정이다. 자동 실행 정책은 별도 범위다.
      required: [transferAssetId, transferDay]
      properties:
        transferAssetId: { type: string, format: uuid }
        transferDay: { type: integer, minimum: 1, maximum: 31 }
    AssetInput:
      type: object
      required: [assetTypeId, ownershipScope, name, openedOn, openingBalanceWon]
      properties:
        assetTypeId:
          type: string
          format: uuid
          description: 현재 가계부의 고정 시스템 자산 종류 ID
        ownershipScope: { $ref: "#/components/schemas/AssetOwnershipScope" }
        ownerMemberId:
          oneOf:
            - { type: string, format: uuid }
            - { type: "null" }
        name: { type: string, minLength: 1, maxLength: 100 }
        openedOn: { type: string, format: date }
        memo:
          oneOf:
            - { type: string, maxLength: 1000 }
            - { type: "null" }
        openingBalanceWon: { type: integer, format: int64 }
        cardSettings:
          description: CREDIT_CARD 유형에서는 필수, 그 외 유형에서는 null이어야 한다.
          oneOf:
            - { $ref: "#/components/schemas/CardSettingsInput" }
            - { type: "null" }
        debitCardSettings:
          description: DEBIT_CARD 유형에서는 필수, 그 외 유형에서는 null이어야 한다.
          oneOf:
            - { $ref: "#/components/schemas/DebitCardSettingsInput" }
            - { type: "null" }
        savingsSettings:
          description: SAVINGS 유형에서는 필수, 그 외 유형에서는 null이어야 한다.
          oneOf:
            - { $ref: "#/components/schemas/SavingsSettingsInput" }
            - { type: "null" }
    CreateAssetRequest:
      allOf:
        - { $ref: "#/components/schemas/AssetInput" }
    UpdateAssetRequest:
      allOf:
        - { $ref: "#/components/schemas/AssetInput" }
        - type: object
          required: [expectedVersion, reassignTransactionsToNewOwner]
          properties:
            expectedVersion: { type: integer, format: int64, minimum: 0 }
            reassignTransactionsToNewOwner:
              type: boolean
              description: PERSONAL owner가 실제 변경될 때만 true를 허용하며 그 외에는 false다.
    CardSettings:
      type: object
      required: [statementClosingDay, paymentDay, paymentMonthOffset, settlementAssetId, autoSettlementEnabled]
      properties:
        statementClosingDay: { type: integer, minimum: 1, maximum: 31 }
        paymentDay: { type: integer, minimum: 1, maximum: 31 }
        paymentMonthOffset: { type: integer, minimum: 0, maximum: 2 }
        settlementAssetId:
          oneOf:
            - { type: string, format: uuid }
            - { type: "null" }
        autoSettlementEnabled: { type: boolean }
    DebitCardSettings:
      allOf:
        - { $ref: "#/components/schemas/DebitCardSettingsInput" }
    SavingsSettings:
      allOf:
        - { $ref: "#/components/schemas/SavingsSettingsInput" }
    Asset:
      type: object
      required:
        - assetId
        - assetTypeId
        - systemCode
        - assetTypeName
        - behavior
        - paymentSourceCapable
        - ownershipScope
        - ownerMemberId
        - name
        - openedOn
        - memo
        - openingBalanceWon
        - currentBalanceWon
        - currentMonthCardPaymentDueWon
        - nextMonthCardPaymentDueWon
        - status
        - archivedAt
        - version
        - cardSettings
        - debitCardSettings
        - savingsSettings
      properties:
        assetId: { type: string, format: uuid }
        assetTypeId: { type: string, format: uuid }
        systemCode:
          type: string
          enum: [CASH, BANK, CREDIT_CARD, DEBIT_CARD, SAVINGS, INVESTMENT, OVERDRAFT, LOAN, INSURANCE, OTHER]
        assetTypeName: { type: string }
        behavior: { $ref: "#/components/schemas/AssetBehavior" }
        paymentSourceCapable: { type: boolean }
        ownershipScope: { $ref: "#/components/schemas/AssetOwnershipScope" }
        ownerMemberId:
          oneOf:
            - { type: string, format: uuid }
            - { type: "null" }
        name: { type: string }
        openedOn: { type: string, format: date }
        memo:
          oneOf:
            - { type: string }
            - { type: "null" }
        openingBalanceWon: { type: integer, format: int64 }
        currentBalanceWon: { type: integer, format: int64 }
        currentMonthCardPaymentDueWon:
          type: integer
          format: int64
          minimum: 0
          description: Asia/Seoul 기준 현재 달에 결제 예정인 이 자산의 미결제 카드 명세 합계이며 카드가 아니면 0이다.
        nextMonthCardPaymentDueWon:
          type: integer
          format: int64
          minimum: 0
          description: Asia/Seoul 기준 다음 달에 결제 예정인 이 자산의 미결제 카드 명세 합계이며 카드가 아니면 0이다.
        status:
          type: string
          enum: [ACTIVE, ARCHIVED]
        archivedAt:
          oneOf:
            - { type: string, format: date-time }
            - { type: "null" }
        version: { type: integer, format: int64 }
        cardSettings:
          oneOf:
            - { $ref: "#/components/schemas/CardSettings" }
            - { type: "null" }
        debitCardSettings:
          oneOf:
            - { $ref: "#/components/schemas/DebitCardSettings" }
            - { type: "null" }
        savingsSettings:
          oneOf:
            - { $ref: "#/components/schemas/SavingsSettings" }
            - { type: "null" }
    AssetRemovalDisposition:
      type: string
      enum: [DELETE, ARCHIVE]
    AssetRemovalBlockingLink:
      type: object
      required: [kind, assetId, assetName]
      properties:
        kind:
          type: string
          enum: [CREDIT_CARD_SETTLEMENT, DEBIT_CARD_PAYMENT, SAVINGS_TRANSFER, CARD_PAYMENT_SCHEDULE]
        assetId: { type: string, format: uuid }
        assetName: { type: string }
    AssetRemovalPreview:
      type: object
      required:
        - assetId
        - name
        - disposition
        - currentBalanceWon
        - historyTransactionCount
        - unpaidCardStatementCount
        - blockingLinks
        - expectedVersion
        - previewToken
      properties:
        assetId: { type: string, format: uuid }
        name: { type: string }
        disposition: { $ref: "#/components/schemas/AssetRemovalDisposition" }
        currentBalanceWon: { type: integer, format: int64 }
        historyTransactionCount: { type: integer, format: int64, minimum: 0 }
        unpaidCardStatementCount: { type: integer, format: int64, minimum: 0 }
        blockingLinks:
          type: array
          items: { $ref: "#/components/schemas/AssetRemovalBlockingLink" }
        expectedVersion: { type: integer, format: int64, minimum: 0 }
        previewToken: { type: string, minLength: 64, maxLength: 64 }
    AssetRemovalResult:
      type: object
      required: [assetId, name, disposition, currentBalanceWon, removedAt]
      properties:
        assetId: { type: string, format: uuid }
        name: { type: string }
        disposition:
          type: string
          enum: [DELETED, ARCHIVED]
        currentBalanceWon: { type: integer, format: int64 }
        removedAt: { type: string, format: date-time }
...
    Problem:
      type: object
      required: [status, title, detail, errorCode, correlationId, timestamp, fieldErrors]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string, format: uri-reference }
        errorCode: { type: string }
        correlationId: { type: string }
        timestamp: { type: string, format: date-time }
        fieldErrors:
          type: array
          items:
            type: object
            additionalProperties: true
  responses:
    BadRequest:
      description: 입력값 또는 token 오류
      content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" } } }
    Unauthorized:
      description: 인증 실패
      content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" } } }
...
    Conflict:
      description: 현재 상태와 충돌하는 요청
      content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" } } }
    NotFound:
      description: 현재 사용자가 접근할 수 있는 리소스가 없음
      content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" } } }
    PreconditionFailed:
      description: 편집 시작 후 다른 세션이 리소스를 수정함
      content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" } } }